refactor(cfd-2d): Decompose turbulence/boundary_conditions/mod.rs — extract manager module
Part of ATLAS-ARCH-007.
